Vfs and underlying fs changes
Each filesystem is now registered (added to a list of filesystems).
They are identified by their name: devfs, tarfs, etc.

So all filesystems need to do is register and mount itself to a specific
path to be initialized. This allows for a shorter initialization for
each filesystem and is definitely an improvement over the previous design.